add random.h
