+++ graphbolt/src/csc_sampling_graph.cc
@@ -6,6 +6,7 @@
 
 #include <graphbolt/csc_sampling_graph.h>
 #include <graphbolt/serialize.h>
+#include <graphbolt/random.h>
 #include <torch/torch.h>
 
 #include <tuple>
